requirements.txt
fastapi==0.98.0
uvicorn[standard]==0.22.0
pydantic==1.10.9
python-dateutil==2.8.2
numpy==1.26.4
python-dotenv==1.0.0
pytest==7.4.0
requests>=2.28.0
pyzmq>=25.0.0
redis>=4.5.0
